Did a global seach for random() and changed it to the ansi standard rand().
Made a new file sensTestFull that does a sensitivity test of all the parameters
using the ranges specified in the .param file.  Updated makefile to it compiles when you do
make all.  To run it you just do sensTestFull fileName numRuns typeNum1 ... 
The output are two files: fileName.Params.sensTest and fileName.Values.sensTest.
Each row corresponds in both files to the corresponding parameter in niwots.param, where a column
in the param file is the value used in the sensitivity, and the columns in the Values file is the value
of the log likelihood.

Made a file niwots_parameters.txt which is just a short description of all the parameters in docs

Replaced turnoverRate to woodTurnoverRate, coarseRootTurnover to coarseRootTurnoverRate,
fineRootTurnover to fineRootTurnoverRate to be more specific.



git-svn-id: svn://cgdm-bishorn.wlclient.ucar.edu/Users/sacks/temporary/sipnet.svn/trunk@13 c88c3319-8f09-0410-a662-fffcf0821eeb