-- Header files on the out files can now be printed with a simple command. At the command line for sipnet and option is -h 1 (h: header on out file, 1 sets it to be true).  The default is zero, so sipnet niwots prints no headers.
-- An "ecosystem-scale girdling" experiment can be done with sipnetGirdle. The
new file niwots.girdle which lists the date that the girdling begins, and
after that any exudates or root growth is set to zero.  All the roots "die"
and enter the soil pool.  This currently can only worked 
be used for the forward version of sipnet, so you
can't run a parameter estimation with girdling.

-- The makefile is adjusted to compile sipnetGirdle

-- Two folders added:
SIPNEToptsIn: code still in development that allows for a separate
model options file to be read in to sipnet instead of having a long list of preprocsseor
directions.

SIPNETparamsChange: code in development that has a more user-friendly parameters file
(fileName.parameters)

The program parametersSort.c  reads in
this file and makes niwots.param, but sorts all the parameters
alphabetically.  Complie this program with the command:
gcc parametersSort.c -o parametersSort

Output should be an alphabetic file for fileName.param

git-svn-id: svn://cgdm-bishorn.wlclient.ucar.edu/Users/sacks/temporary/sipnet.svn/trunk@18 c88c3319-8f09-0410-a662-fffcf0821eeb