Bill Sacks, 3-13-07
Major changes:
	- Modified parameter reading so that order of parameters doesn't matter, and it's okay if some parameters aren't present in the input file. (Note, however, that any spatially-varying parameters must appear in the same order in the spatial parameter file as in the regular parameter file.) C
	- Comments are now allowed in the *.param file (but not in the *.param-spatial file). Any text after a '!' is interpreted as a comment.
	- NUM_PARAMS (in sipnet.c) now determined dynamically, saving you the trouble of having to count all the time!

See below for specific changes.


--- SPATIALPARAMS.C ---

Added and deleted a number of helper functions to account for new mechanisms for handling the parameter structure

Deleted readOneSpatialParam function, replaced it with initializeOneSpatialParam (called explicitly for each parameter) and readSpatialParams (called once to read the whole file)
	- The initializeOneSpatialParam function establishes a connection between a given parameter name and a variable location
	- The readSpatialParams function then makes use of this information to read in all the parameter values
		- Comments can now be included; the comment character is '!'
	- It is okay for a parameter to be missing from the input file, unless it was given the isRequired attribute when calling initializeOneSpatialParam (see changes to sipnet.c:readParamData for more information)

Fixed bug in resetSpatialParams function where it would assign ALL parameters to a random start value if randomReset=true, even non-changeable parameters


--- SIPNET.C ---

Modified readParamData function to use new parameter reading routines
	- For each parameter, specify a boolean value (1 or 0) indicating whether this parameter is required. If it is, the program will terminate if the parameter is missing from the input file. If it's not required, a warning message will be printed, but the run will continue. 
		- I envision that parameters will either be given a value of 1 (always required), or a boolean expression that evaluates to true if and only if the model component using that parameter is activated (see the existing code for examples). Of course, it is legal to provide a value of 0, but you do so at your own (and others') risk, since the error checking is then diminished. 
		- Important note: I got tired of trying to figure out all the appropriate flags after the "qualityWood" parameter, and flagged the rest of them as 0 (optional). If people could go through and flag their own parameters appropriately, that would be a great help. 

I moved a few lines of code to be between #if... #endif directives, to allow more parameters to be optional (these changes shouldn't change the model output at all for any of the options; they just change, e.g., whether some values are set to 0 implicitly or explicitly)

Changed sipnet.c:NUM_PARAMS to be determined dynamically
	- NUM_PARAMS = sizeof(Params)/sizeof(double)
	- This removes the annoying requirement of having to count the number of parameters
	- Note that this includes all parameters, not just the ones read from file

--- UTIL.C ---

Added strcmpIgnoreCase function
      - Replaces use of non-standard strcasecmp

Added stripComment function
      - Strips any trailing comment off line
      - Return value: 1 if line contains only a comment (or only blanks), 0 otherwise


--- TO DO ---

sipnetGirdle doesn't compile: readParamData needs to be changed to match the new readParamData in sipnet.c

I gave up on flagging parameters as required or optional after params.qualityWood (in sipnet.c:readParamData). Right now all remaining parameters are marked as optional. While this is convenient, it is not very safe. If you get a chance, it would be great if people could flag the parameters you've created in a way similar to what I've done for the other parameters (i.e. give a boolean value that evaluates to true if and only if the module using that parameter is active).



git-svn-id: svn://cgdm-bishorn.wlclient.ucar.edu/Users/sacks/temporary/sipnet.svn/trunk@20 c88c3319-8f09-0410-a662-fffcf0821eeb