updated sipnet.c paramchange.c subsetData.c To allow uncertainties for different data types to be read in to sipnet during the estimate proceedure
NEW REQUIREMENT
a file called filename.sigma must be provided.  This file should be in the same format as the filename.dat and filename.valid files and it should contain the errors associated with each of the data types

I have also included old versions of paramchange and the difference function for reference so that I can use these to update the code to allow filename.sigma to be ignored and estimate to use the old method.

git-svn-id: svn://cgdm-bishorn.wlclient.ucar.edu/Users/sacks/temporary/sipnet.svn/branches/moore@80 c88c3319-8f09-0410-a662-fffcf0821eeb